requirements.txt
pytest~=9.0.2
selenium~=4.39.0
allure-pytest
allure-python-commons
webdriver-manager==*
allure-pytest==2.*